requirements.txt
streamlit
requests
speechrecognition
pydub
python-dotenv
python-dotenv
